green-formfix/backend/requirements.txt
fastapi[all]==0.128.1
SQLAlchemy==2.0.48
PyMySQL==1.1.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
uvicorn==0.40.0
python-dotenv==1.1.1
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
tensorflow==2.15.0
tensorflow-metal==1.1.0
scikit-learn
